Switch to using an input wrapper around raw xorg
An input class was made to handle the raw xorg calls that
XDisplay makes. This also makes it easier in the future to
add wayland support, and this lets there be only 1 call to
an update function which means the screen updates at one time
rather than multiple times per iteration.

Signed-off-by: Maxwell Henderson <maxwell.henderson@mailbox.org>